Module renderComments.js is added
